requirements.txt
requests
environs
pydantic
python-telegram-bot
gradio-client